feat: production-grade hardening - thread safety, memory management, error handling
- FixedWindowLimiter: Thread-safe WindowState class with lock
- RedisTokenBucket: Added 1-hour TTL for key expiry
- Redis: Graceful error handling with fallback to in-memory
- SlidingWindow: Periodic stale entry cleanup
- Middleware: Restored middleware chain execution
- All 44 tests passing with zero warnings

+++ DistributedRateLimiter.Tests/RedisHealthCheckTests.cs
@@ -121,6 +121,6 @@ public class RedisHealthCheckTests
         var result = await _healthCheck.CheckHealthAsync(new HealthCheckContext());
 
         Assert.True(result.Data.ContainsKey("redis_failure_count"));
-        Assert.Equal(2, (long)result.Data["redis_failure_count"]);
+        Assert.Equal(2, (int)result.Data["redis_failure_count"]);
     }
 }
